jobs out

+++ skeleton_smash/Commands.cpp
@@ -23,7 +23,7 @@ SmallShell* SmallShell::instance = nullptr;
 
 ///////////////////////////////////////////////  General Functions   ///////////////////////////////////////////////////////////
 
-std::string fileNameOpen(const char* cmd_line)
+std::string fileNameOpen(char* cmd_line)
 {
     std::string fileName = cmd_line;
     if(fileName.find(">>") < fileName.length())
@@ -36,7 +36,7 @@ std::string fileNameOpen(const char* cmd_line)
     }
 }
 
-bool redirection(const char* cmd_line, bool setTimeout)
+bool redirection(char* cmd_line, bool setTimeout)
 {
     string line = cmd_line;
     bool append = false;
@@ -129,10 +129,11 @@ bool _isBackgroundComamnd(const char* cmd_line)
     return str[str.find_last_not_of(WHITESPACE)] == '&';
 }
 
-void _removeBackgroundSign(string cmd_line)
+void _removeBackgroundSign(char* cmd_line)
 {
+    string str(cmd_line);
     // find last character other than spaces
-    unsigned int idx = cmd_line.find_last_not_of(WHITESPACE);
+    unsigned int idx = str.find_last_not_of(WHITESPACE);
     // if all characters are spaces then return
     if (idx == (unsigned int)string::npos)
     {
@@ -145,16 +146,21 @@ void _removeBackgroundSign(string cmd_line)
     // replace the & (background sign) with space and then remove all tailing spaces.
     cmd_line[idx] = ' ';
     // truncate the command line string up to the last non-space character
-    cmd_line[cmd_line.find_last_not_of(WHITESPACE, idx) + 1] = 0;
+    cmd_line[str.find_last_not_of(WHITESPACE, idx) + 1] = '\0';
+    cout << cmd_line << "after" << endl;
 }
 
-int numOfWords(const char* cmd_line, string* argsTable)
+int numOfWords(char* cmd_line, string* argsTable)
 {
-    string line = _trim(cmd_line);
+    char* temp=cmd_line;
+    _removeBackgroundSign(temp);
+    std::string line = _trim(cmd_line);
+    //line = _trim(line);
     int count = 1;
     string cur;
     cur += line[0];
     int index = 0;
+    cout << line << "end" << endl;
     for(int i = 1 ; i < (int)line.length() ; i++)
     {
         if(line[i-1] != ' ' && line[i] == ' ')
@@ -171,7 +177,7 @@ int numOfWords(const char* cmd_line, string* argsTable)
     return count;
 }
 
-void splitByArg(const char* line, char arg, char** result)
+void splitByArg(char* line, char arg, char** result)
 {
     string lineString = _trim(line);
     string arg1, arg2;
@@ -194,9 +200,9 @@ void splitByArg(const char* line, char arg, char** result)
     result[1]= strCopy2;
 }
 
-string findCommand(const char* cmd_line)
+string findCommand(char* cmd_line)
 {
-    string line = cmd_line;
+    string line = _trim(cmd_line);
     string command = "";
     string timeout = "timeout";
     int i = 0;
@@ -218,6 +224,16 @@ string findCommand(const char* cmd_line)
     for ( ; i != (int)line.length() && isalpha(line[i]) ; i++)
         command += line[i];
     return command;
+}
+
+bool isNum(std::string s)
+{
+    for (int i = 0 ; i < (int)s.length() ; i++)
+    {
+        if (!isdigit(s[i]))
+            return false;
+    }
+    return true;
 }
 
 ///////////////////////////////////////////////////////   Commands   ///////////////////////////////////////////////////////////
@@ -265,7 +281,7 @@ int JobsList::countJobs() const
     return counter;
 }
 
-void JobsList::addJob(const char* cmd_line, pid_t pid, bool isStopped)
+void JobsList::addJob(char* cmd_line, pid_t pid, bool isStopped)
 {
     if (isStopped)
     {
@@ -320,12 +336,13 @@ void JobsList::removeFinishedJobs()
         if (BGround[i-1])
         {
             int wait_res = waitpid(BGround[i - 1]->getPid(), nullptr, WNOHANG);
-
+            //cout << "pid : " << BGround[i - 1]->getPid() << endl;
             if (wait_res == -1) {
                 perror("smash error: waitpid failed");
             }
             if (wait_res == BGround[i - 1]->getPid()) {
-                //delete BGround[i - 1];
+                //cout << wait_res << " delete job" << endl;
+                delete BGround[i - 1];
                 BGround[i - 1] = nullptr;
             }
         }
@@ -458,7 +475,7 @@ pid_t JobsList::JobEntry::getPid() const
     return pid;
 }
 
-const char* JobsList::JobEntry::getCmdLine()
+char* JobsList::JobEntry::getCmdLine()
 {
     return cmdLine;
 }
@@ -476,7 +493,7 @@ SmallShell::~SmallShell()
     jobsList->killAllJobs();
 }
 
-void SmallShell::changeName(const char* newName)
+void SmallShell::changeName(string newName)
 {
     namePrompt = newName;
 }
@@ -486,7 +503,7 @@ pid_t SmallShell::getSmashPid() const
     return smashPid;
 }
 
-void SmallShell::addCD(const char* dir)
+void SmallShell::addCD(char* dir)
 {
     curCD = dir;
 }
@@ -509,13 +526,17 @@ JobsList* SmallShell::getJobs()
 /**
 * Creates and returns a pointer to Command class which matches the given command line (cmd_line)
 */
-Command* SmallShell::CreateCommand(const char* cmd_line)
+Command* SmallShell::CreateCommand(char* cmd_line)
 {
-    string cmd_s = _trim(string(cmd_line));
+    char * temp=cmd_line;
+    _removeBackgroundSign(temp);
+    string cmd_s = _trim(string(temp));
     bool runInBack = false;
     if (_isBackgroundComamnd(cmd_line))
+    {
+        cout << "run " << endl;
         runInBack = true;
-    _removeBackgroundSign(cmd_s);
+    }
     bool isChild  = false;
     bool redirectionHappened = false;
     Command* cmd = nullptr;
@@ -561,7 +582,7 @@ Command* SmallShell::CreateCommand(const char* cmd_line)
     return cmd;
 }
 
-bool SmallShell::forkExtrenal(bool setTimeout, bool runInBack, const char* cmd_line)
+bool SmallShell::forkExtrenal(bool setTimeout, bool runInBack, char* cmd_line)
 {
     pid_t child_pid = fork();
     if(child_pid < 0)
@@ -601,10 +622,10 @@ bool SmallShell::forkExtrenal(bool setTimeout, bool runInBack, const char* cmd_l
     return false;
 }
 
-Command* SmallShell::BuiltIn(const char* cmd_line)
+Command* SmallShell::BuiltIn(char* cmd_line)
 {
     string firstWord = _trim(cmd_line);
-    firstWord = findCommand(firstWord.c_str());
+    firstWord = findCommand(cmd_line);
     if (firstWord == "chprompt")
         return new ChpromptCommand(cmd_line);
     else if (firstWord == "showpid")
@@ -650,10 +671,10 @@ std::vector<Timeout_obj*> SmallShell::getAlarmed()
     return timeout;
 }
 
-void SmallShell::executeCommand(const char *cmd_line) {
+void SmallShell::executeCommand(char *cmd_line) {
     char* cmd = new char[strlen(cmd_line)] ;
     strcpy(cmd, cmd_line);
-    const char* cmdLine = cmd;
+    char* cmdLine = cmd;
     Command* command = CreateCommand(cmdLine);
     if(command == nullptr)
     {
@@ -676,7 +697,7 @@ void ChpromptCommand::execute()
     }
     else
     {
-        SmallShell::getInstance().changeName(args[1].c_str());
+        SmallShell::getInstance().changeName(args[1]);
     }
 }
 
@@ -741,16 +762,14 @@ void ForegroundCommand::execute()
             int pid = stoi(firstArg);
             if (pid <= 0 )
             {
-                perror("1");
-                string error = "smash error:fg:job-id " + to_string(pid) + " does not exist";
+                string error = "smash error: fg: job-id " + to_string(pid) + " does not exist";
                 perror(error.c_str());
                 return;
             }
             JobsList::JobEntry* job = jobs->getJobById(pid);
             if (!job)
             {
-                perror("2");
-                string error = "smash error:fg:job-id " + to_string(pid) + " does not exist";
+                string error = "smash error: fg: job-id " + to_string(pid) + " does not exist";
                 perror(error.c_str());
                 return;
             }
@@ -769,7 +788,7 @@ void ForegroundCommand::execute()
         }
         catch (exception &e)
         {
-            perror("smash error: fg:invalid arguments");
+            perror("smash error: fg: invalid arguments");
         }
     }
     else if (argsCount == 1)
@@ -795,7 +814,7 @@ void ForegroundCommand::execute()
     }
     else
     {
-        perror("smash error:fg:invalid arguments");
+        perror("smash error: fg: invalid arguments");
     }
 }
 
@@ -812,33 +831,33 @@ void BackgroundCommand::execute()
             int pid = stoi(firstArg);
             if (pid <= 0 )
             {
-                string error = "smash error:bg:job-id" + to_string(pid) + "does not exist";
+                string error = "smash error: bg: job-id " + to_string(pid) + " does not exist";
                 perror(error.c_str());
                 return;
             }
             JobsList::JobEntry* job = jobs->getJobById(pid);
             if (!job)
             {
-                string error = "smash error:bg:job-id " + to_string(pid) + " does not exist";
+                string error = "smash error: bg: job-id " + to_string(pid) + " does not exist";
                 perror(error.c_str());
                 return;
             }
             else if (job->getStat() != stopped)
             {
-                string error = "smash error:bg:job-id " + to_string(pid) + " is already running in the background";
+                string error = "smash error: bg: job-id " + to_string(pid) + " is already running in the background";
                 perror(error.c_str());
                 return;
             }
             jobs->moveToBG(job);
             cout << job->getCmdLine() << " : " << job->getPid() << endl;
-            if(kill(pid, SIGCONT)==-1)
+            if(kill(pid, SIGCONT) == -1)
             {
                 perror("smash error: kill failed");
             }
         }
         catch (exception &e)
         {
-            perror("smash error: bg:invalid arguments");
+            perror("smash error: bg: invalid arguments");
         }
     }
     else if (argsCount == 1)
@@ -846,19 +865,19 @@ void BackgroundCommand::execute()
         JobsList::JobEntry* job = jobs->getLastStoppedJob();
         if (!job)
         {
-            perror("smash error: bg:there is no stopped jobs to resume");
+            perror("smash error: bg: there is no stopped jobs to resume");
             return;
         }
         jobs->moveToBG(job);
         cout << job->getCmdLine() << " : " << job->getPid() << endl;
-        if(kill(job->getPid(), SIGCONT)==-1)
+        if(kill(job->getPid(), SIGCONT) == -1)
         {
             perror("smash error: kill failed");
         }
     }
     else
     {
-        perror("smash error:bg:invalid arguments");
+        perror("smash error: bg: invalid arguments");
     }
 }
 
@@ -874,6 +893,7 @@ void SimpleCommand::execute()
     std::string argsTable[22];
     int argsCnt = numOfWords(cmdLine, argsTable);
     bool exists = _isBackgroundComamnd(cmdLine);
+    argsTable[argsCnt-1] = _trim(argsTable[argsCnt-1]);
     if (exists)
     {
         if (argsTable[argsCnt-1] != "&")
@@ -887,7 +907,8 @@ void SimpleCommand::execute()
         argsTable[i]= _trim(argsTable[i]);
         char* strCopy = new char[argsTable[i].size()+1];
         std::strcpy(strCopy,argsTable[i].c_str());
-        argv[i]= strCopy;
+        cout << strCopy << "     " << i << endl;
+        argv[i] = strCopy;
     }
     argv[argsCnt] = nullptr;
     execvp(argsTable[0].c_str(), argv);
@@ -931,7 +952,7 @@ void ComplexCommand::execute()
         exit(errno);
 }
 
-PipeCommand::PipeCommand(const char* cmd_line): Command(cmd_line)
+PipeCommand::PipeCommand(char* cmd_line): Command(cmd_line)
 {
     curCommand= nullptr;
     string cmd_s = _trim(string(cmdLine));
@@ -1061,7 +1082,7 @@ void SetcoreCommand::execute()
     }
 }
 
-bool is_file_exist(const char *fileName)
+bool is_file_exist(char *fileName)
 {
     std::ifstream infile(fileName);
     return infile.good();
@@ -1183,22 +1204,38 @@ void KillCommand::execute()
 {
     SmallShell::getInstance().getJobs()->removeFinishedJobs();
     std::string arg[22];
-    if(numOfWords(cmdLine,arg)!=3 || arg[1].length()>3 || arg[1].length()==1)
+    std::string sigNum;
+    if(numOfWords(cmdLine,arg) != 3 || arg[1].length() > 3 || arg[1].length() == 1)
     {
         perror("smash error: kill: invalid arguments");
+        return;
     }
     try
     {
-        std::string sigNum = arg[1].substr(1, arg[1].length() - 1);
+        if (arg[1][0] == '-')
+        {
+            sigNum = arg[1].substr(1, arg[1].length() - 1);
+        }
+        if (!isNum(sigNum) || !isNum(arg[2]))
+        {
+            perror("smash error: kill: invalid arguments");
+            return;
+        }
         JobsList::JobEntry *job = SmallShell::getInstance().getJobs()->getJobById(stoi(arg[2]));
         if (job)
         {
-            cout << "signal number " + sigNum + " was sent to pid " + to_string(job->getPid()) << endl;
+            if (waitpid(job->getPid(), nullptr, WNOHANG) != 0)
+            {
+                std::string errorMessage = "smash error: kill: job-id " + arg[2] + " does not exist";
+                perror(errorMessage.c_str());
+                return;
+            }
             if(kill(job->getPid(), stoi(sigNum)) == -1)
             {
                 perror("smash error: kill failed");
+                return;
             }
-            return;
+            cout << "signal number " + sigNum + " was sent to pid " + to_string(job->getPid()) << endl;
         }
         else
         {
